mas bima ini tolong dianuin crit

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -48,7 +48,7 @@ return new class extends Migration
             $table->string('ip_address', 45)->nullable();
             $table->text('user_agent')->nullable();
             $table->longText('payload');
-            $table->integer('last_activity')->index();
+            $table->bigInteger('last_activity')->index();
         });
     }
