+++ database/migrations/2026_05_09_072848_create_asuransi_kendaraans_table.php
@@ -32,7 +32,7 @@
                 $table->date('tgl_berakhir');
 
                 // Tambahan
-                $table->integer('durasi_bulan');
+                $table->bigInteger('durasi_bulan');
                 $table->decimal('biaya', 15, 2);
 
                 $table->string('bukti_bayar')->nullable();
