+++ database/migrations/2026_07_08_000006_create_sosmedps_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
             $table->string('channel');
             $table->string('utm_source');
             $table->string('utm_campaign');
-            $table->integer('klik')->default(0);
-            $table->integer('konversi')->default(0);
+            $table->bigInteger('klik')->default(0);
+            $table->bigInteger('konversi')->default(0);
             $table->decimal('total_biaya', 15, 2)->default(0);
             $table->decimal('total_penjualan', 15, 2)->default(0);
             $table->decimal('roi', 10, 2)->default(0);
