+++ database/migrations/2026_07_12_000001_create_reminder_service_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->date('tanggal_mulai');
 
             // Interval reminder
-            $table->integer('interval_nilai')->default(1);
+            $table->bigInteger('interval_nilai')->default(1);
             $table->enum('interval_satuan', ['hari', 'minggu', 'bulan', 'tahun'])->default('bulan');
 
             // Tanggal jatuh tempo (dihitung otomatis: tanggal_mulai + interval)
